Replaced data with data sql by removing data and renaming dataSQL to data dataSQLTest.scala renamed to dataTest.scala
