+++ src/main/scala/PrivateHome/data.scala
@@ -1,47 +1,179 @@
 package PrivateHome
 
+import PrivateHome.Devices.MHz.mhzSwitch
+import PrivateHome.Devices.MQTT.mqttSwitch
 import PrivateHome.Devices.Switch
+import scalikejdbc._
 
-import scala.xml._
 
 object data {
-
   /**
-   * This variable returns a List containing all ids.
+   * An map containing all settings
    */
-  val IDs: List[String] = {
-    val xml = XML.load("src/main/scala/PrivateHome/id.xml")
-    val prep: NodeSeq = xml \\ "id"
-    val result = for (i <- 0 until prep.length) yield prep(i).text
-    result.toList
-  }
-
-  /**
-   * This variable returns a map which contains the id as the key and the Switch-object for this id.
-   */
-  val devices: Map[String, Switch] = {
-    val xml: Elem = XML.load("src/main/scala/PrivateHome/devices.xml")
-    val prep: NodeSeq = xml \ "switch"
-    val data: Seq[(String, Switch)] = for (i <- prep) yield (i \ "@id").text -> Switch(i)
-    data.toMap
-  }
-
-  /**
-   * This variable returns an map, which contains the MHz-Code(systemCode+unitCode) as the key and the id for this code.
-   */
-  val mhzID: Map[String, String] = {
-    val xml: Elem = XML.load("src/main/scala/PrivateHome/devices.xml")
-    val prep: NodeSeq = xml \ "switch"
-    val data: Seq[(String, String)] = for (i <- prep) yield ((i \ "systemCode").text + (i \ "unitCode").text) -> (i \ "@id").text
-    data.toMap
-  }
-
   val settings: Map[String, Int] = Map(("port", 2888), ("ip", 565))
 
-  def idTest(id:String):Unit ={
+
+  GlobalSettings.loggingSQLAndTime = LoggingSQLAndTimeSettings(
+    enabled = false,
+    singleLineMode = false,
+    printUnprocessedStackTrace = false,
+    stackTraceDepth = 15,
+    logLevel = Symbol("debug"),
+    warningEnabled = true,
+    warningThresholdMillis = 3000L,
+    warningLogLevel = Symbol("warn")
+  )
+
+
+  Class.forName("org.h2.Driver")
+  ConnectionPool.singleton("jdbc:h2:./daten/hello", "user", "pass")
+
+  implicit val session: AutoSession.type = AutoSession
+  var devices: Map[String, Switch] = Map()
+  var mhzId: Map[String, String] = Map()
+
+  /**
+   * Generates the Table strucktures for the Database
+   */
+  def create(): Unit = {
+
+    // Drops old Tables to Delet Data and make it possible to regenerate them
+    sql"""DROP TABLE IF EXISTS `Mhz`;
+         DROP TABLE IF EXISTS `Devices`;""".execute().apply()
+
+
+    val devices =
+      sql"""
+                    CREATE TABLE `Devices` (
+                    `id` varchar(5) NOT NULL,
+                    `name` varchar(64) NOT NULL,
+                    `type` varchar(16) NOT NULL,
+                    `state` decimal(5,4) NOT NULL,
+                    `keepState` boolean NOT NULL,
+                    PRIMARY KEY (`id`))
+         """.execute.apply()
+
+
+    sql"""
+           CREATE TABLE `Mhz` (
+           `id` varchar(5) NOT NULL,
+           `systemcode` varchar(5) NOT NULL,
+           `unitcode` varchar(5) NOT NULL,
+           PRIMARY KEY (`id`),
+           CONSTRAINT `Mhz_ibfk_1` FOREIGN KEY (`id`) REFERENCES `Devices` (`id`))
+         """.execute().apply()
+
+
+    // insert initial data
+
+
+  }
+
+  /**
+   * Fills the devices Map with all Devices from the DB and turns all Switches with keepState on
+   */
+  def fillDevices(): Unit = {
+    val m = Device.syntax("m")
+    withSQL {
+      select.from(Device as m)
+    }.map(rs => Device(rs)).list().apply().foreach(d => {
+      d.switchtype match {
+        case "MQTT" => devices += ((d.id, mqttSwitch(d.id, keepStatus = d.keepState)))
+          if (d.keepState) {
+            devices(d.id).on(d.state)
+          }
+        case "433Mhz" =>
+          val m = Mhz.syntax("m")
+          val data = withSQL {
+            select.from(Mhz as m).where.eq(m.id, d.id)
+          }.map(rs => Mhz(rs)).single().apply().get
+          devices += ((d.id, mhzSwitch(d.id, d.keepState, data.systemcode, data.unitcode)))
+          mhzId += ((data.systemcode + data.unitcode,d.id))
+          if (d.keepState) {
+            devices(d.id).on(d.state)
+          }
+      }
+
+    })
+
+  }
+
+  /**
+   * Adds an Switch to the DB witch all needed under Tables
+   *
+   * @param device The device that should be added
+   */
+  def addDevice(device: Switch): Unit = {
+    var wrongclass = new IllegalArgumentException("""Can not add Switch ID:{} because switch of unknown type {} has no save definition""".format(device.id, device.getClass))
+    withSQL {
+      insertInto(Device).values(device.id, "", device.switchtype, if (device.keepStatus) device.Status else 0, device.keepStatus)
+    }.update.apply
+    device match {
+      case device: mhzSwitch => withSQL {
+        insertInto(Mhz).values(device.id, device.systemCode, device.unitCode)
+      }.update().apply()
+        mhzId += ((device.systemCode + device.unitCode, device.id))
+      case _: mqttSwitch =>
+      case _ => throw wrongclass
+    }
+    devices = devices.concat(Map((device.id, device)))
+  }
+  //ToDo: add an settings XML
+  //ToDo: add support for Settingschanges
+
+  def idTest(id: String): Unit = {
     if (id.length != 5) throw new IllegalArgumentException("""Length of ID is not 5""")
     if (!id.matches("[-_a-zA-Z0-9]{5}")) throw new IllegalArgumentException("""ID Contains not Allowed Characters""")
-    if (data.IDs.contains(id))throw new IllegalArgumentException("""ID is already used""")
+    if (devices.contains(id)) throw new IllegalArgumentException("""ID is already used""")
+  }
+
+  /**
+   * An better access to the devices Map
+   *
+   * @param deviceID The ID of the switch you want
+   * @return the Switch object the you requested
+   */
+  def getDevice(deviceID: String): Switch = {
+    devices(deviceID)
+  }
+
+  /**
+   * Message class for devices Table
+   *
+   * @param id         ID of the Device in the Format [0-9a-Z] five character long
+   * @param name       The name of the Device any String lenght in the Table 64 character
+   * @param switchtype A String identification of the Switch type max length 16 character
+   * @param state      an foatingpoint representation of the State when keepState is true 4 decimalpoints long
+   * @param keepState  Boolean that indicates if the State should be restored at turn on
+   */
+  case class Device(id: String, name: String, switchtype: String, state: Float, keepState: Boolean)
+
+  /**
+   * Message class for Mhz Table only needed when Switchtype is MQTT
+   *
+   * @param id         ID of the Device in the Format [0-9a-Z] five character long
+   * @param systemcode Systemcode for the Mhzreciever
+   * @param unitcode   Unitcode for the Mhzreciever
+   */
+  case class Mhz(id: String, systemcode: String, unitcode: String)
+
+  /**
+   * Syntaxsupport Object for devices table
+   */
+  object Device extends SQLSyntaxSupport[Device] {
+    override val tableName = "devices"
+
+    def apply(rs: WrappedResultSet) = new Device(
+      rs.string("id"), rs.string("name"), rs.string("type"), rs.float("state"), rs.boolean("keepState"))
+  }
+
+  /**
+   * Syntaxsupport Object for Mhz table
+   */
+  object Mhz extends SQLSyntaxSupport[Mhz] {
+    override val tableName = "Mhz"
+
+    def apply(rs: WrappedResultSet): Mhz = new Mhz(rs.string("id"), rs.string("systemcode"), rs.string("unitcode"))
   }
 
 }
