+++ src/main/scala/PrivateHome/main.scala
@@ -5,7 +5,7 @@ import PrivateHome.UI.GUI.gui
 object main {
   def main(args: Array[String]): Unit = {
 
-    dataSQL
+    data
     gui
   }
 }
